requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.4.1
    # via django
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.7
    # via requests
django==3.2.9
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-filter
    #   djangorestframework
django-extensions==3.1.5
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
djangorestframework==3.12.4
    # via -r requirements.in
google-i18n-address==2.5.0
    # via -r requirements.in
idna==3.3
    # via requests
pycountry==20.7.3
    # via -r requirements.in
pytz==2021.3
    # via django
pyyaml==6.0
    # via -r requirements.in
requests==2.26.0
    # via google-i18n-address
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via -r requirements.in
urllib3==1.26.7
    # via requests
